feat: add hook to get col ratio text color
